refactor(OC): migrate request token handling to Typescript
